finish 1 vs N

+++ database/migrations/2022_04_11_083556_add_user_id_to_posts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->bigInteger('user_id')->unsigned();
+            $table->Integer('user_id')->unsigned();
             $table->foreign('user_id')->references('user_id')->on('users')->onDelete('cascade');
 
         });
